+++ MultiPanelControl/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0.1] - [08 Dec 2015]
+### New
+- Included page selector in MultiPanel designer
+- Included MultiPanelPage parent selector
+
 ## [1.0.0.0] - [17 Jun 2009]
 ### Changed
 - Initial release